refactor: move variables from svc_override.go to override.go
